Implemented commandBuffer and commandHandler, started work on command processing

+++ Ricardo_OS/Ricardo_OS/src/Comms/packets.h
@@ -20,11 +20,10 @@ struct tx_packet_t{
 
 struct rx_packet_t{
 
-    uint32_t command;
+    uint8_t command;
     
-
-
 };
+
